requirements.txt
fastapi
python-socketio
uvicorn
redis